pyproject.toml
[project]
name = "intent-pilot"
version = "0.1.0"
description = "what can be said can be automated"
authors = [
    {name = "gitlost-murali-askui", email = "murali.kondragunta@askui.com"},
]
dependencies = [
    "python-dotenv>=1.0.1",
    "Pillow>=10.2.0",
    "requests>=2.31.0",
    "openai>=1.12.0",
    "PyAutoGUI>=0.9.54",
    "plyer>=2.1.0",
    "prompt-toolkit>=3.0.43",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.scripts]
intent = "python -m intent_pilot.run.app"

[project.scripts]
intent = "intent_pilot.run.app:main"
